nueva clase ExpNumerica

+++ TSP-TS/src/main/java/Tabu/Tabu.java
@@ -143,7 +143,7 @@ public class Tabu {
         return solucion;
     }
 
-    public double getDuracion(){
+    public long getDuracion(){
         return this.duracion;
     }
